Use git-cherry-pick instead of pygit2
Our simplistic implementation using pygit2 suffers from several
shortcomings:

- We are only handling one commit at a time, so sequencing (abort,
  continue, skip) does not work properly: Abort does not rewind the
  previously applied commits, continue does not apply more than the
  current commit, and skip simply fails.

- The commit message includes the list of conflicting files.

There is another shortcoming that is not fixed by this change: When the
user aborts a cherry-pick in progress, a previously created branch is
not deleted. The correct behavior in this case also requires some more
thought. (Maybe we should simply refrain from creating branches for the
user?)